+++ lazy_test.go
@@ -12,7 +12,7 @@ func TestLazy(t *testing.T) {
 	t.Run("Lazy1", func(t *testing.T) {
 		var result string
 		compute := func(s string) { result = s }
-		Lazy1(compute)(Return("string"))
+		Lazy1(compute)(Thunk("string"))
 		assert.That(t, assert.Equal("string", result))
 	})
 	t.Run("Lazy2", func(t *testing.T) {
@@ -20,7 +20,7 @@ func TestLazy(t *testing.T) {
 		compute := func(s1, s2 string) {
 			result = s1 + s2
 		}
-		Lazy2(compute)(Return("abc"), Return("def"))
+		Lazy2(compute)(Thunk("abc"), Thunk("def"))
 		assert.That(t, assert.Equal("abcdef", result))
 	})
 	t.Run("Lazy3", func(t *testing.T) {
@@ -28,7 +28,7 @@ func TestLazy(t *testing.T) {
 		compute := func(s1, s2 string, i int) {
 			result = s1 + s2 + strconv.FormatInt(int64(i), 16)
 		}
-		Lazy3(compute)(Return("abc"), Return("def"), Return(123))
+		Lazy3(compute)(Thunk("abc"), Thunk("def"), Thunk(123))
 		assert.That(t, assert.Equal("abcdef7b", result))
 	})
 	t.Run("Lazy4", func(t *testing.T) {
@@ -36,7 +36,7 @@ func TestLazy(t *testing.T) {
 		compute := func(s1, s2 string, i int, b bool) {
 			result = s1 + s2 + strconv.FormatInt(int64(i), 16) + strconv.FormatBool(b)
 		}
-		Lazy4(compute)(Return("abc"), Return("def"), Return(123), Return(true))
+		Lazy4(compute)(Thunk("abc"), Thunk("def"), Thunk(123), Thunk(true))
 		assert.That(t, assert.Equal("abcdef7btrue", result))
 	})
 	t.Run("LazyAll0", func(t *testing.T) {
@@ -44,7 +44,7 @@ func TestLazy(t *testing.T) {
 		compute := func(a ...string) {
 			result = strings.Join(a, ", ")
 		}
-		Lazy1S(compute)(ReturnS("abc", "def"))
+		Lazy1S(compute)(ThunkS("abc", "def"))
 		assert.That(t, assert.Equal("abc, def", result))
 	})
 }
@@ -55,35 +55,35 @@ func TestLazyR(t *testing.T) {
 			return s + s
 		}
 		assert.That(t, assert.Equal("stringstring",
-			Lazy1R(compute)(Return("string"))))
+			Lazy1R(compute)(Thunk("string"))))
 	})
 	t.Run("Lazy2R", func(t *testing.T) {
 		compute := func(s1, s2 string) string {
 			return s1 + s2
 		}
 		assert.That(t, assert.Equal("abcdef",
-			Lazy2R(compute)(Return("abc"), Return("def"))))
+			Lazy2R(compute)(Thunk("abc"), Thunk("def"))))
 	})
 	t.Run("Lazy3R", func(t *testing.T) {
 		compute := func(s1, s2 string, i int) string {
 			return s1 + s2 + strconv.FormatInt(int64(i), 16)
 		}
 		assert.That(t, assert.Equal("abcdef7b",
-			Lazy3R(compute)(Return("abc"), Return("def"), Return(123))))
+			Lazy3R(compute)(Thunk("abc"), Thunk("def"), Thunk(123))))
 	})
 	t.Run("Lazy4R", func(t *testing.T) {
 		compute := func(s1, s2 string, i int, b bool) string {
 			return s1 + s2 + strconv.FormatInt(int64(i), 16) + strconv.FormatBool(b)
 		}
 		assert.That(t, assert.Equal("abcdef7btrue",
-			Lazy4R(compute)(Return("abc"), Return("def"), Return(123), Return(true))))
+			Lazy4R(compute)(Thunk("abc"), Thunk("def"), Thunk(123), Thunk(true))))
 	})
 	t.Run("LazyAll", func(t *testing.T) {
 		compute := func(a ...string) string {
 			return strings.Join(a, ", ")
 		}
 		assert.That(t, assert.Equal("abc, def",
-			Lazy1SR(compute)(ReturnS("abc", "def"))))
+			Lazy1SR(compute)(ThunkS("abc", "def"))))
 	})
 }
 
@@ -93,34 +93,34 @@ func TestLazyR2(t *testing.T) {
 			return s + s, true
 		}
 		assert.That(t, assert.True(DropFirstOf2(
-			Lazy1R2(compute)(Return("string")))))
+			Lazy1R2(compute)(Thunk("string")))))
 	})
 	t.Run("LazyTwo2", func(t *testing.T) {
 		compute := func(s1, s2 string) (string, bool) {
 			return s1 + s2, true
 		}
 		assert.That(t, assert.True(DropFirstOf2(
-			Lazy2R2(compute)(Return("abc"), Return("def")))))
+			Lazy2R2(compute)(Thunk("abc"), Thunk("def")))))
 	})
 	t.Run("LazyThree2", func(t *testing.T) {
 		compute := func(s1, s2 string, i int) (string, bool) {
 			return s1 + s2 + strconv.FormatInt(int64(i), 16), true
 		}
 		assert.That(t, assert.True(DropFirstOf2(
-			Lazy3R2(compute)(Return("abc"), Return("def"), Return(123)))))
+			Lazy3R2(compute)(Thunk("abc"), Thunk("def"), Thunk(123)))))
 	})
 	t.Run("LazyFour2", func(t *testing.T) {
 		compute := func(s1, s2 string, i int, b bool) (string, bool) {
 			return s1 + s2 + strconv.FormatInt(int64(i), 16) + strconv.FormatBool(b), true
 		}
 		assert.That(t, assert.True(DropFirstOf2(
-			Lazy4R2(compute)(Return("abc"), Return("def"), Return(123), Return(true)))))
+			Lazy4R2(compute)(Thunk("abc"), Thunk("def"), Thunk(123), Thunk(true)))))
 	})
 	t.Run("LazyAll2", func(t *testing.T) {
 		compute := func(a ...string) (string, bool) {
 			return strings.Join(a, ", "), true
 		}
 		assert.That(t, assert.True(DropFirstOf2(
-			Lazy1SR2(compute)(ReturnS("abc", "def")))))
+			Lazy1SR2(compute)(ThunkS("abc", "def")))))
 	})
 }