+++ wrap_test.go
@@ -11,7 +11,7 @@ import (
 func TestWrap(t *testing.T) {
 	assert.That(t,
 		assert.Equal("A,B,C,DE",
-			Wrap(Wrap(Wrap(Pass(
+			Wrap(Wrap(Wrap(Return(
 				strings.ToUpper,
 			), strings.Fields,
 			), slices.Compact[[]string],
@@ -19,7 +19,7 @@ func TestWrap(t *testing.T) {
 			)(" a a a b c c c c de")),
 
 		assert.Equal(true,
-			Wrap(Wrap(Wrap(Pass(
+			Wrap(Wrap(Wrap(Return(
 				strings.ToUpper,
 			), strings.Fields,
 			), slices.Compact[[]string],