+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-### Added
+### Changed
 
 - Generated functions up to 5 parameters.
+
+### Added
+
 - `Adapt*` to adapt a function to the given signature.
 - `Lazy*` for lazy evaluation of function arguments.
 - `Head*` to make partial currying.
@@ -16,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - `seq` package, that provides some sequence processing functions.
 - `Signature*` and `FnSig*` to compute function parameters and return types without using reflection.
 - `Eq` and `Not` predicates.
+- `Return*` to return the passed argument.
+- `Thunk*` makes a thunk function that returns the passed argument.
 
 ## [0.0.1] - Initial release